feat: add "mask support" for reading & composition (#96)
This PR is updating libCZI to latest version, which includes support for
"mask" (c.f. https://github.com/ZEISS/libczi/pull/153).
The new functionality of "mask-aware tile-composition" is made available
to pylibczirw by introducing a `ReaderOptions` object which optionally
can be passed to `open_czi`. Mask-aware composition is therefore an
opt-in function at this point (motivation is to limit risk and ensure
backwards compatibility).